osutil: refactoring mountinfo code following PR review
My original refactoring of the mountinfo logic introduced
an inconsistency into the code, so that os-agnostic code
contain Linux-related logic. This inconsistencey was
eliminated.
